Task/day09 (#8)
* Day 9 pt 1: relative parameters & update relativeBase opcode.
TODO: memory expansion

* Encapsulate Opcode array into inner class Program and add capability to expand "memory" to address locations beyond initial program size.

* Day 9 pt 1 - need to fix 203 error.

* Completed Day 9 parts 1 & 2! :-)

+++ src/main/java/thebergers/adventofcode2019/day07/Day07Part2.java
@@ -15,7 +15,7 @@ public class Day07Part2 {
 				.setProgram(program)
 				.setUseFeedback(true)
 				.build();
-		Integer maxThrust = tsc.calculateMaxThrust();
+		Long maxThrust = tsc.calculateMaxThrust();
 		System.out.format("Max thrust signal = %d%n", maxThrust);
 	}
 }
